chore(desk-tool): move part implementations to `_parts`
