+++ packages/@sanity/desk-tool/src/_parts/base-document-badges-resolver.ts
@@ -1,0 +1,1 @@
+export {default} from './base-document-badges'
