--- packages/@sanity/desk-tool/src/badges/defaultResolveDocumentBadges.ts
@@ -1,1 +1,0 @@
-export {default} from './defaultDocumentBadges'
